franka/requirements.txt
absl-py==2.1.0
actionlib==1.14.1
agentlace @ git+https://github.com/youliangtan/agentlace.git@cf2c337c5e3694cdbfc14831b239bd657bc4894d
angles==1.9.14
attrs==25.3.0
blinker==1.9.0
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.11
catkin-pkg==1.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.8
cloudpickle==3.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
decorator==4.4.2
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
dm-tree==0.1.9
docstring_parser==0.17.0
docutils==0.21.2
dynamic-reconfigure==1.7.4
easyhid==0.0.10
einops==0.8.0
etils==1.11.0
evdev==1.7.1
Farama-Notifications==0.0.4
Flask==3.1.0
flax==0.10.2
fsspec==2024.12.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.1
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.17
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.14.0
hidapi==0.14.0.post4
humanize==4.11.0
idna==3.10
image-geometry==1.16.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
interactive-markers==1.12.1
itsdangerous==2.2.0
jax==0.4.35
jax-cuda12-pjrt==0.4.35
jax-cuda12-plugin==0.4.35
jaxlib==0.4.34
Jinja2==3.1.5
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
keyboard==0.13.5
laser-geometry==1.6.8
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
message-filters==1.17.0
ml_dtypes==0.5.0
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
moviepy==1.0.3
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.6.0.74
nvidia-cufft-cu12==11.3.0.4
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.6.85
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.0
pillow==11.3.0
proglog==0.1.12
protobuf==5.29.2
pycparser==2.22
Pygments==2.18.0
pymodbus==2.5.3
pynput==1.7.7
pyparsing==3.2.1
pyquaternion==0.9.9
pyrealsense2==2.54.2.5684
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-qt-binding==0.4.5
python-xlib==0.33
PyYAML==6.0.2
pyzmq==26.2.0
qt-dotgraph==0.4.3
qt-gui==0.4.3
qt-gui-cpp==0.4.3
qt-gui-py-common==0.4.3
requests==2.32.3
resource-retriever==1.12.9
rich==13.9.4
rosbag==1.17.0
rosboost-cfg==1.15.9
rosclean==1.15.9
roscreate==1.15.9
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.9
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.9
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.5.1
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.9
roswtf==1.17.0
rqt-action==0.4.10
rqt-bag==0.5.2
rqt-bag-plugins==0.5.2
rqt-console==0.4.13
rqt-dep==0.4.13
rqt-graph==0.4.15
rqt-gui==0.5.4
rqt-gui-py==0.5.4
rqt-image-view==0.4.18
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.11
rqt-plot==0.4.15
rqt-pose-view==0.5.12
rqt-publisher==0.4.11
rqt-py-common==0.5.4
rqt-py-console==0.4.11
rqt-reconfigure==0.5.6
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.13
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.1
rqt-service-caller==0.4.11
rqt-srv==0.4.10
rqt-tf-tree==0.6.4
rqt-top==0.4.11
rqt-topic==0.4.14
rqt-web==0.4.11
rqt_launch==0.4.9
rqt_nav_view==0.5.7
rqt_shell==0.4.11
rviz==1.14.25
scipy==1.15.0
sensor-msgs==1.13.1
-e git+https://github.com/rail-berkeley/hil-serl.git@7d17d13560d85abffbd45facec17c4f9189c29c0#egg=serl_launcher&subdirectory=serl_launcher
-e git+https://github.com/rail-berkeley/hil-serl.git@7d17d13560d85abffbd45facec17c4f9189c29c0#egg=serl_robot_infra&subdirectory=serl_robot_infra
shtab==1.7.2
simplejson==3.19.3
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.6
srdfdom==0.6.4
svgwrite==1.4.3
tensorstore==0.1.71
tf==1.13.3
tf-conversions==1.13.3
tf2-geometry-msgs==0.7.8
tf2-kdl==0.7.8
tf2-py==0.7.8
tf2-ros==0.7.8
toolz==1.0.0
topic-tools==1.17.0
tqdm==4.67.1
Tree==0.2.4
typeguard==4.4.4
typing==3.7.4.3
typing_extensions==4.15.0
urdfdom-py==0.4.6
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xacro==1.14.20
zipp==3.21.0
zmq==0.0.0
